% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erpPCA-package.R
\docType{data}
\name{dien2005}
\alias{dien2005}
\title{Data from tutorial on PCA applications to Event-Related Potentials data by Dien and Frishkoff.}
\format{A data frame with 10 columns and 20 rows}
\usage{
dien2005
}
\description{
This dataset is used by Dien and Frishkoff on their book chapter ``Principal Components
  Analysis of ERP Data'' (see reference below) for a step by step tutorial.
}
\section{Variables}{


\itemize{
\item \code{Ss}: Subject codes.
\item \code{Cs}: Condition codes.
\item \code{t1-t6}: Time samples 1 to 6.
}
}
\references{
Dien, J., & Frishkoff, C. A. (2005). Principal Components
  Analysis of ERP Data. In Handy, T. (Ed.) Event-related Potentials: A
  Methods Handbook (pp. 189-207). Cambridge, MA: MIT Press.
}

