% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erpPCA.R
\name{erpPCA}
\alias{erpPCA}
\title{Performs unrestricted, unstandardized covariance-based PCA with Varimax
rotation (cf. Kayser J, Tenke CE, Clin Neurophysiol, 2003, 114:2307-25)}
\source{
\url{http://psychophysiology.cpmc.columbia.edu/mmedia/Kayser2003a/Appendix.html}
}
\usage{
erpPCA(X)
}
\arguments{
\item{X}{A matrix. This is the data matrix to be decomposed into its
  principal components. For a temporal decomposition (ie., latent factors
  will correspond to time courses), the matrix should be organized
  with ERP waveforms of each sensor in rows (cases) and time samples in
  columns (variables). For spatial decomposition (ie., latent factors
  will correspond to scalp topographies), the matrix should be organized
  with time samples in rows (cases) and sensors in columns (variables).}
}
\value{
This function returns a list containing the unrotated principal
  component loadings, the Varimax rotated factor loadings in a
  variables-by-factors matrix, the factor scores obtained for the rotated
  factors in a cases-by-factors matrix. It also outputs the Eingenvalues
  of the covariance matrix and the their explained variance before and after
  rotation.
}
\description{
This function implements a PCA decomposition of a data matrix
based on its unstandardized covariance matrix, followed by a Varimax rotation
of all of its principal components, in an attempt to find simpler latent
factors.
}
\details{
This workflow of matrix decomposition followed by Varimax rotation
has been advocated for electrophysiological studies on humans using the
Event-Related Potential (ERP) technique by J\"{u}rgen Kayser and Craig E.
Tenke. These authors suggest that very good results can be obtained by using
(1) unstandardized covariance matrices of the original data matrix followed
by (2) a Varimax rotation of all of the principal components (what they call
an ``unrestricted'' rotation). The arguments are laid out in a series of
papers (see references below)

This is function is a R port of the original MATLAB function erpPCA.m created
by J\"{u}rgen Kayser. According to the original documentation, it implements
``the PCA agorithms used by BMDP-4M (Dixon, 1992) and SPSS 10.0 FACTOR''. The
original MATLAB code can be freely downloaded at
http://psychophysiology.cpmc.columbia.edu/mmedia/Kayser2003a/Appendix.html
}
\examples{
## using the iris dataset and principal components analysis
erpPCA(iris[, 2:4])
}
\references{
Kayser, J., & Tenke, C. E. (2003). Optimizing PCA methodology for
  ERP component identification and measurement: theoretical rationale and
  empirical evaluation. Clinical neurophysiology, 114(12), 2307-2325.

Kayser, J., & Tenke, C. E. (2005). Trusting in or breaking with
  convention: towards a renaissance of principal components analysis in
  electrophysiology. Clinical Neurophysiology, 116(8), 1747-1753.

Kayser, J., & Tenke, C. E. (2006). Consensus on PCA for ERP data,
  and sensibility of unrestricted solutions. Clinical Neurophysiology,
  117(3), 703-707.
}
\seealso{
\code{\link[stats]{princomp}}, \code{\link[stats]{prcomp}},
  \code{\link[stats]{factanal}} for other basic implementations of principal
  component analysis and factor analysis.
}

