% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erpPCA.R
\name{erpPCA}
\alias{erpPCA}
\title{Performs unrestricted, unstandardized covariance-based PCA followed by
  Varimax rotation}
\source{
\url{http://psychophysiology.cpmc.columbia.edu/mmedia/Kayser2003a/Appendix.html}
}
\usage{
erpPCA(X, method = c("eigen", "svd"), rank.method = c("original", "qr"),
  rotation.fun = c("original", "R"))
}
\arguments{
\item{X}{A matrix. This is the data matrix to be decomposed into its
  principal components.}

\item{method}{A string. Method to use for finding principal components of
  data matrix. Default is \code{eigen} to match the original MATLAB function,
  but \code{svd} is generally recommended for numerical accuracy. Both should
  give the same results.}

\item{rank.method}{A string. Method to use to estimate the rank of the
  principal component matrix. Default is \code{original}, to match the
  original MATLAB function. Option \code{qr} may be more computationally
  efficient, but will give different results from the original method.}

\item{rotation.fun}{A string. Function to use for Varimax rotation. Default
  is \code{original}, which calls the function \code{\link{DoVarimax4M}} from this package,
  to match the original MATLAB function. Option \code{R} will call the base
  function \code{varimax}. The outputs of these choices will probably be
  numerically different.}
}
\value{
This function returns a list containing the unrotated principal
  component loadings, the Varimax rotated factor loadings in a
  variables-by-factors matrix, the factor scores obtained for the rotated
  factors in a cases-by-factors matrix. It also outputs the Eingenvalues
  of the covariance matrix and the their explained variance before and after
  rotation.

\item{Unrotated}{Matrix containing the unrotated principal components
loadings.}

\item{Rotated}{Matrix containing the varimax rotated factor loadings}

\item{Factor.Scores}{Matrix containing the factor scores of the varimax
rotated factor loadings.}

\item{Variance}{Matrix containing the Eigenvalues and explained variance
before and afte rotation.}
}
\description{
This function implements a PCA decomposition of a data matrix
based on its unstandardized covariance matrix, followed by a Varimax rotation
of all of its principal components, in an attempt to find simpler latent
factors.
}
\details{
This workflow of matrix decomposition followed by Varimax rotation
  has been advocated for electrophysiological studies on humans using the
  Event-Related Potential (ERP) technique by Jurgen Kayser and Craig E.
  Tenke. These authors suggest that very good results can be obtained by
  using (1) unstandardized covariance matrices of the original data matrix
  followed by (2) a Varimax rotation of all of the principal components (what
  they call an ``unrestricted'' rotation). The arguments are laid out in a
  series of papers (see references below)

This function is a R port of the original MATLAB function erpPCA.m created
by Jurgen Kayser. According to the original documentation, it implements
``the PCA agorithms used by BMDP-4M (Dixon, 1992) and SPSS 10.0 FACTOR''. The
original MATLAB code can be freely downloaded at
\url{http://psychophysiology.cpmc.columbia.edu/mmedia/Kayser2003a/Appendix.html}
}
\section{Temporal vs Spatial PCA}{

For a temporal decomposition (ie., latent factors will correspond to time
courses), the matrix \code{X} should be organized with ERP waveforms of each
sensor in rows (cases) and time samples in columns (variables). For spatial
decomposition (ie., latent factors will correspond to scalp topographies),
the matrix \code{X} should be organized with time samples in rows (cases) and
sensors in columns (variables).
}
\examples{
## using the iris dataset and principal components analysis
erpPCA(iris[, 2:4])

## Using the tutorial dataset from Dien and Frishkoff (2005)
X <- dien2005[, 3:8]

## Matrix decomposition methods (eigen and svd) should give identical results
## no matter what the matrix rank estimation method is used.
A <- erpPCA(X, method = "eigen", rank.method = "original",
  rotation.fun = "original")
B <- erpPCA(X, method = "svd", rank.method = "original",
  rotation.fun = "original")
all.equal(A, B, check.attributes = FALSE)

C <- erpPCA(X, method = "eigen", rank.method = "qr",
  rotation.fun = "original")
D <- erpPCA(X, method = "svd", rank.method = "qr", rotation.fun = "original")
all.equal(C, D, check.attributes = FALSE)

## The varimax rotation options will give numerically different results,
## however. The magnitude of this difference needs to be gauged on a case
## by case basis.
E <- erpPCA(X, method = "svd", rank.method = "original", rotation.fun = "R")
all.equal(B, E, check.attributes = FALSE)
}
\references{
Kayser, J., & Tenke, C. E. (2003). Optimizing PCA methodology for
  ERP component identification and measurement: theoretical rationale and
  empirical evaluation. Clinical neurophysiology, 114(12), 2307-2325.

Kayser, J., & Tenke, C. E. (2005). Trusting in or breaking with
  convention: towards a renaissance of principal components analysis in
  electrophysiology. Clinical Neurophysiology, 116(8), 1747-1753.

Kayser, J., & Tenke, C. E. (2006). Consensus on PCA for ERP data,
  and sensibility of unrestricted solutions. Clinical Neurophysiology,
  117(3), 703-707.
}
\seealso{
\code{\link[stats]{princomp}}, \code{\link[stats]{prcomp}},
  \code{\link[stats]{factanal}} from package \code{stats} for other basic
  implementations of principal component analysis and factor analysis.
}

